format staff tree view

+++ MainWindow.cpp
@@ -7,6 +7,7 @@
 #include <QString>
 #include <QModelIndex>
 #include <QMessageBox>
+#include "StringFormatter.h"
 
 using namespace std;
 
@@ -42,10 +43,16 @@ MainWindow::~MainWindow() {
     delete ui;
 }
 
-void MainWindow::areButtonsEnable(bool active)
+void MainWindow::arePatientButtonsEnable(bool active)
 {
     ui->deletePatientButton->setEnabled(active);
     ui->editPatientButton->setEnabled(active);
+}
+
+void MainWindow::areStaffButtonsEnable(bool active)
+{
+    ui->deleteStaffButton->setEnabled(active);
+    ui->editStaffButton->setEnabled(active);
 }
 
 void MainWindow::setupPatientTab()
@@ -61,7 +68,7 @@ void MainWindow::setupPatientTab()
     ui->patientsTableView->setEditTriggers(QTableView::NoEditTriggers);
     ui->patientsTableView->show();
 
-    areButtonsEnable(false);
+    arePatientButtonsEnable(false);
 }
 
 void MainWindow::setupStaffTab()
@@ -71,9 +78,11 @@ void MainWindow::setupStaffTab()
     QList<StaffType> typeList = _staffTypeConnector->getAll();
 
     QStandardItem * rootNode = _typesModel->invisibleRootItem();
-    _typesModel->setColumnCount(2);
+    _typesModel->setColumnCount(4);
     _typesModel->setHeaderData(0,Qt::Horizontal,"Prénom");
     _typesModel->setHeaderData(1,Qt::Horizontal,"Nom");
+    _typesModel->setHeaderData(2,Qt::Horizontal,"Consultation");
+    _typesModel->setHeaderData(3,Qt::Horizontal,"Durée");
 
     for (int i = 0; i < typeList.size(); ++i) {
        QStandardItem * item =  new QStandardItem(typeList.at(i).getName().c_str());
@@ -82,18 +91,42 @@ void MainWindow::setupStaffTab()
     rootNode->appendRows(_typeItemsList);
 
     for (int i = 0; i < staffList.size(); i++) {
+        Staff staff = staffList.at(i);
+        QList<Patient> patients = _patientConnector->getAll(staff.getId());
+        staff.setPatientsConsult(patients);
         for (int j = 0; j < _typeItemsList.size(); ++j) {
             if (staffList.at(i).getTypeId() == typeList.at(j).getId()) {
                 QList<QStandardItem*> items;
-                items.append(new QStandardItem(staffList.at(i).getFirstName().c_str()));
-                items.append(new QStandardItem(staffList.at(i).getLastName().c_str()));
+                items.append(new QStandardItem(staff.getFirstName().c_str()));
+                items.append(new QStandardItem(staff.getLastName().c_str()));
+                QList<Patient> consultations = staff.getPatientsConsult();
+                QString dates = "", durations = "";
+                for (int k = 0; k < consultations.size(); ++k) {
+                    dates += ("- " + StringFormatter::toFrDate(consultations.at(k).getConsultationDate()) + (k+1 < consultations.size() ? "\n" : ""));
+                    durations += ("- " + StringFormatter::toReadableDuration(consultations.at(k).getDuration()) + (k+1 < consultations.size() ? "\n" : ""));
+                }
+                items.append(new QStandardItem(dates));
+                items.append(new QStandardItem(durations));
+
                 _typeItemsList.at(j)->appendRow(items);
             }
         }
+        // display satff
+        staff.display();
     }
     ui->treeView->setModel(_typesModel);
     ui->treeView->setEditTriggers(QTreeView::NoEditTriggers);
     ui->treeView->setColumnWidth(0, 200);
+    ui->treeView->setColumnWidth(1, 150);
+    ui->treeView->setColumnWidth(2, 150);
+    ui->treeView->expandAll();
+
+    areStaffButtonsEnable(false);
+
+    //selection changes shall trigger a slot
+    QItemSelectionModel *selectionModel= ui->treeView->selectionModel();
+    connect(selectionModel, SIGNAL(selectionChanged (const QItemSelection &, const QItemSelection &)),
+            this, SLOT(selectionChangedSlot(const QItemSelection &, const QItemSelection &)));
 }
 
 void MainWindow::on_addStaffPushButton_clicked()
@@ -180,7 +213,7 @@ void MainWindow::on_searchButton_clicked()
 {
     qDebug() << ui->searchTextBox->text();
     _patientConnector->searchFilterModel(ui->searchTextBox->text());
-    areButtonsEnable(false);
+    arePatientButtonsEnable(false);
 
 }
 
@@ -246,7 +279,7 @@ void MainWindow::on_editPatientButton_clicked()
                 _consultConnector->insert(consultToAdd[i]);
             }
 
-            // Delete
+            // DeleteS
             for (int i = 0; i < consultToDelete.size(); i++) {
                 _consultConnector->suppr(consultToDelete[i]._id);
             }
@@ -258,7 +291,7 @@ void MainWindow::on_editPatientButton_clicked()
 
 void MainWindow::on_patientsTableView_pressed(const QModelIndex &index)
 {
-    areButtonsEnable(true);
+    arePatientButtonsEnable(true);
     if (index.isValid() && (_patientClickedIndex == -1 || _patientClickedIndex != index.row())) {
         _patientClickedIndex = index.row();
     }
@@ -277,23 +310,17 @@ void MainWindow::on_addPatientPushButton_clicked()
 
 void MainWindow::on_deletePatientButton_clicked()
 {
-    // Delete consult involving this patient
-    QList<Consult> consultToDelete = _consultConnector->getPatientConsult(_patientsModel->record(_patientClickedIndex).value("Id").toInt());
+    QMessageBox::StandardButton reply;
+      reply = QMessageBox::question(this, "Attention", "Voulez-vous vraiment supprimer ce patient ?",QMessageBox::Yes|QMessageBox::No);
+      if (reply == QMessageBox::Yes) {
+        ui->patientsTableView->hideRow(_patientClickedIndex);
+        ui->patientsTableView->setModel(_patientsModel);
+        _patientConnector->deleteRecord(_patientClickedIndex);
+        ui->statusBar->showMessage("Patient supprimé");
+      }
+}
 
-    for (int i = 0; i < consultToDelete.size(); i++) {
-        qDebug() << "delete : " << consultToDelete[i]._id << ", " << consultToDelete[i]._idPatient << ", " << consultToDelete[i]._idRessource;
-        _consultConnector->suppr(consultToDelete[i]._id);
-    }
-
-    qDebug() << "remainning consult";
-
-    QList<Consult> test = _consultConnector->getAll();
-    for (int i = 0; i < test.size(); i++) {
-        qDebug() << test[i]._id << ", " << test[i]._idPatient << ", " << test[i]._idRessource;
-    }
-
-    // delete the patient
-    ui->patientsTableView->hideRow(_patientClickedIndex);
-    ui->patientsTableView->setModel(_patientsModel);
-    _patientConnector->deleteRecord(_patientClickedIndex);
+void MainWindow::selectionChangedSlot(const QItemSelection &newSelection, const QItemSelection &oldSelection)
+{
+    areStaffButtonsEnable(true);
 }
